+++ node/parallel/src/chain_spec/parallel.rs
@@ -16,10 +16,12 @@ use cumulus_primitives_core::ParaId;
 use heiko_runtime::pallet_loans::{InterestRateModel, JumpModel, Market, MarketState};
 use hex_literal::hex;
 use parallel_runtime::{
-    opaque::SessionKeys, BalancesConfig, CollatorSelectionConfig, CouncilConfig, DemocracyConfig,
-    ElectionsConfig, GenesisConfig, LiquidStakingAgentMembershipConfig, LiquidStakingConfig,
-    LoansConfig, OracleMembershipConfig, ParachainInfoConfig, SessionConfig, SudoConfig,
-    SystemConfig, TechnicalCommitteeConfig, TokensConfig, WASM_BINARY,
+    opaque::SessionKeys, pallet_nominee_election::NomineeCoefficients, BalancesConfig,
+    CollatorSelectionConfig, CouncilConfig, DemocracyConfig, ElectionsConfig, GenesisConfig,
+    LiquidStakingAgentMembershipConfig, LiquidStakingConfig, LoansConfig, NomineeElectionConfig,
+    OracleMembershipConfig, ParachainInfoConfig, SessionConfig, SudoConfig, SystemConfig,
+    TechnicalCommitteeConfig, TokensConfig, ValidatorFeedersMembershipConfig, VestingConfig,
+    WASM_BINARY,
 };
 use primitives::{network::NetworkType, *};
 use sc_service::ChainType;
@@ -56,7 +58,6 @@ pub fn development_config(id: ParaId) -> ChainSpec {
                 vec![get_account_id_from_seed::<sr25519::Public>("Ferdie")],
                 vec![
                     get_account_id_from_seed::<sr25519::Public>("Dave"),
-                    get_account_id_from_seed::<sr25519::Public>("Eve"),
                     get_account_id_from_seed::<sr25519::Public>("Alice//stash"),
                     get_account_id_from_seed::<sr25519::Public>("Bob//stash"),
                     get_account_id_from_seed::<sr25519::Public>("Charlie//stash"),
@@ -64,7 +65,8 @@ pub fn development_config(id: ParaId) -> ChainSpec {
                     get_account_id_from_seed::<sr25519::Public>("Eve//stash"),
                     get_account_id_from_seed::<sr25519::Public>("Ferdie//stash"),
                 ],
-                // Multisig account combined by Alice, Bob and Charile, ss58 prefix is 42
+                vec![get_account_id_from_seed::<sr25519::Public>("Eve")],
+                // Multisig account combined by Alice, Bob and Charile, ss58 prefix is 42
                 vec!["5DjYJStmdZ2rcqXbXGX7TW85JsrW6uG4y9MUcLq2BoPMpRA7"
                     .parse()
                     .unwrap()],
@@ -148,7 +150,10 @@ pub fn local_testnet_config(id: ParaId) -> ChainSpec {
                         .parse()
                         .unwrap(),
                 ],
-                // Parallel team accounts, ss58 prefix is 42
+                vec!["5FjH9a7RQmihmb7i4UzbNmecjPm9WVLyoJHfsixkrLGEKwsJ"
+                    .parse()
+                    .unwrap()],
+                // Parallel team accounts, ss58 prefix is 42
                 vec!["5HHMY7e8UAqR5ZaHGaQnRW5EDR8dP7QpAyjeBu6V7vdXxxbf"
                     .parse()
                     .unwrap()],
@@ -171,9 +176,15 @@ fn testnet_genesis(
     invulnerables: Vec<(AccountId, AuraId)>,
     oracle_accounts: Vec<AccountId>,
     endowed_accounts: Vec<AccountId>,
+    validator_feeders: Vec<AccountId>,
     liquid_staking_agents: Vec<AccountId>,
     id: ParaId,
 ) -> GenesisConfig {
+    let vesting_list: Vec<(AccountId, BlockNumber, BlockNumber, u32, Balance)> =
+        serde_json::from_str(include_str!(
+            "../../../../resources/parallel-vesting-PARA.json"
+        ))
+        .unwrap();
     GenesisConfig {
         system: SystemConfig {
             code: WASM_BINARY
@@ -185,6 +196,7 @@ fn testnet_genesis(
             balances: {
                 let mut endowed_accounts = endowed_accounts.clone();
                 endowed_accounts.extend_from_slice(&oracle_accounts);
+                endowed_accounts.extend_from_slice(&validator_feeders);
                 endowed_accounts.extend(
                     invulnerables
                         .iter()
@@ -317,6 +329,20 @@ fn testnet_genesis(
         liquid_staking_agent_membership: LiquidStakingAgentMembershipConfig {
             members: liquid_staking_agents,
             phantom: Default::default(),
+        },
+        validator_feeders_membership: ValidatorFeedersMembershipConfig {
+            members: validator_feeders,
+            phantom: Default::default(),
+        },
+        vesting: VestingConfig {
+            vesting: vesting_list,
+        },
+        nominee_election: NomineeElectionConfig {
+            coefficients: NomineeCoefficients {
+                crf: 100,
+                nf: 1000,
+                epf: 10,
+            },
         },
     }
 }