+++ pallets/loans/src/lib.rs
@@ -52,6 +52,7 @@ mod market;
 mod mock;
 mod rate_model;
 mod tests;
+
 pub mod weights;
 
 /// Container for borrow balance information
@@ -1109,6 +1110,7 @@ impl<T: Config> Pallet<T> {
         collateral_currency_id: CurrencyId,
     ) -> DispatchResult {
         Self::ensure_currency(&liquidate_currency_id)?;
+        Self::ensure_currency(&collateral_currency_id)?;
 
         let market = Self::market(&liquidate_currency_id)?;